mi2gc/requirements.txt
fastapi==0.115.6
pydantic==2.10.4
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic-settings~=2.7.0
httpx==0.28.1
